test: concentrate phase 6 integration owners
Move mock-heavy facade, health, security, and MCP contracts into focused unit owners, keep integration on real archive and repository behavior, and centralize shared adversarial and MCP test helpers.
