+++ tests/infra/mcp.py
@@ -1,18 +1,72 @@
-"""Shared fixtures for integration tests."""
+"""Shared MCP test helpers and surface contracts."""
 
 from __future__ import annotations
 
+import asyncio
 from datetime import datetime, timezone
 from unittest.mock import AsyncMock, MagicMock
 
-import pytest
-
 from polylogue.lib.models import Conversation, Message
 
+EXPECTED_TOOL_NAMES = {
+    "search",
+    "list_conversations",
+    "get_conversation",
+    "stats",
+    "add_tag",
+    "remove_tag",
+    "list_tags",
+    "get_metadata",
+    "set_metadata",
+    "delete_metadata",
+    "delete_conversation",
+    "get_conversation_summary",
+    "get_session_tree",
+    "get_stats_by",
+    "health_check",
+    "rebuild_index",
+    "update_index",
+    "export_conversation",
+}
 
-@pytest.fixture
-def mock_repo():
-    """Create a mock ConversationRepository."""
+EXPECTED_RESOURCE_URIS = {
+    "polylogue://stats",
+    "polylogue://conversations",
+    "polylogue://tags",
+    "polylogue://health",
+}
+
+EXPECTED_RESOURCE_TEMPLATE_URIS = {
+    "polylogue://conversation/{conv_id}",
+}
+
+EXPECTED_PROMPT_NAMES = {
+    "analyze_errors",
+    "summarize_week",
+    "extract_code",
+    "compare_conversations",
+    "extract_patterns",
+}
+
+
+def invoke_surface(fn, /, *args, **kwargs):
+    """Call an MCP surface whether it is sync or async."""
+    result = fn(*args, **kwargs)
+    if asyncio.iscoroutine(result):
+        return asyncio.run(result)
+    return result
+
+
+async def invoke_surface_async(fn, /, *args, **kwargs):
+    """Await an MCP surface from async tests."""
+    result = fn(*args, **kwargs)
+    if asyncio.iscoroutine(result):
+        return await result
+    return result
+
+
+def make_repo_mock() -> MagicMock:
+    """Create a repository mock with async methods used by MCP surfaces."""
     repo = MagicMock()
     repo.list = AsyncMock(return_value=[])
     repo.search = AsyncMock(return_value=[])
@@ -21,6 +75,7 @@ def mock_repo():
     repo.resolve_id = AsyncMock(return_value=None)
     repo.get_archive_stats = AsyncMock(return_value=MagicMock())
     repo.get_summary = AsyncMock(return_value=None)
+    repo.get_conversation_stats = AsyncMock(return_value={})
     repo.get_session_tree = AsyncMock(return_value=[])
     repo.get_stats_by = AsyncMock(return_value={})
     repo.add_tag = AsyncMock(return_value=None)
@@ -30,20 +85,13 @@ def mock_repo():
     repo.update_metadata = AsyncMock(return_value=None)
     repo.delete_metadata = AsyncMock(return_value=None)
     repo.delete_conversation = AsyncMock(return_value=False)
+    repo.backend = MagicMock()
     return repo
 
 
 def make_mock_filter(results=None, **method_overrides):
-    """Create a pre-configured mock ConversationFilter.
-
-    Args:
-        results: List of conversations to return from .list()
-        **method_overrides: Set side_effect for any method
-
-    Returns:
-        Configured MagicMock filter instance with chaining support.
-    """
-    f = MagicMock()
+    """Create a chaining-capable ConversationFilter mock."""
+    filt = MagicMock()
     for method in (
         "contains",
         "exclude_text",
@@ -64,20 +112,19 @@ def make_mock_filter(results=None, **method_overrides):
         "before",
         "tags",
     ):
-        getattr(f, method).return_value = f
-    f.list = AsyncMock(return_value=results or [])
+        getattr(filt, method).return_value = filt
+    filt.list = AsyncMock(return_value=results or [])
     for method_name, override_value in method_overrides.items():
-        method = getattr(f, method_name)
+        method = getattr(filt, method_name)
         if isinstance(override_value, Exception):
             method.side_effect = override_value
         else:
             method.return_value = override_value
-    return f
+    return filt
 
 
-@pytest.fixture
-def simple_conversation():
-    """Create a simple conversation for testing."""
+def make_simple_conversation() -> Conversation:
+    """Return a representative conversation for MCP surface tests."""
     return Conversation(
         id="test:conv-123",
         provider="chatgpt",
