--- tests/integration/test_mcp_exports.py
@@ -1,228 +1,0 @@
-"""MCP server export and completeness tests — export tool, new resources, new prompts, safe_call, tool inventory."""
-
-from __future__ import annotations
-
-import asyncio
-import json
-from unittest.mock import AsyncMock, MagicMock, patch
-
-import pytest
-
-from tests.integration.conftest import make_mock_filter
-
-
-def _invoke_surface(fn, /, *args, **kwargs):
-    """Call an MCP surface whether it is sync or async."""
-    result = fn(*args, **kwargs)
-    if asyncio.iscoroutine(result):
-        return asyncio.run(result)
-    return result
-
-
-async def _invoke_surface_async(fn, /, *args, **kwargs):
-    """Await an MCP surface from async tests."""
-    result = fn(*args, **kwargs)
-    if asyncio.iscoroutine(result):
-        return await result
-    return result
-
-
-def _make_repo_mock() -> MagicMock:
-    """Create a repository mock with async methods for MCP handlers."""
-    repo = MagicMock()
-    repo.view = AsyncMock(return_value=None)
-    repo.get = AsyncMock(return_value=None)
-    repo.list_tags = AsyncMock(return_value={})
-    return repo
-
-# =============================================================================
-# Tier 5: Export Tool Tests
-# =============================================================================
-
-
-class TestExportConversationTool:
-    """Tests for export_conversation tool."""
-
-    def test_export_markdown(self, simple_conversation):
-        """export_conversation with markdown format calls _format_conversation."""
-        from polylogue.mcp.server import _build_server
-
-        with patch("polylogue.mcp.server._get_repo") as mock_get_repo:
-            with patch("polylogue.lib.formatting.format_conversation") as mock_format:
-                mock_repo = _make_repo_mock()
-                mock_repo.view.return_value = simple_conversation
-                mock_get_repo.return_value = mock_repo
-                mock_format.return_value = "# Test Conversation\n\nFormatted content"
-
-                server = _build_server()
-                result = _invoke_surface(server._tool_manager._tools["export_conversation"].fn,
-                    id="test:conv-123", format="markdown"
-                )
-
-                assert "Test Conversation" in result
-                mock_format.assert_called_once()
-                call_args = mock_format.call_args
-                assert call_args[0][0] == simple_conversation
-                assert call_args[0][1] == "markdown"
-
-    def test_export_not_found(self):
-        """export_conversation returns error if conversation not found."""
-        from polylogue.mcp.server import _build_server
-
-        with patch("polylogue.mcp.server._get_repo") as mock_get_repo:
-            mock_repo = _make_repo_mock()
-            mock_repo.view.return_value = None
-            mock_get_repo.return_value = mock_repo
-
-            server = _build_server()
-            result = _invoke_surface(server._tool_manager._tools["export_conversation"].fn, id="nonexistent")
-
-            parsed = json.loads(result)
-            assert "error" in parsed
-            assert "not found" in parsed["error"].lower()
-
-    def test_export_invalid_format(self, simple_conversation):
-        """export_conversation falls back to markdown for invalid format."""
-        from polylogue.mcp.server import _build_server
-
-        with patch("polylogue.mcp.server._get_repo") as mock_get_repo:
-            with patch("polylogue.lib.formatting.format_conversation") as mock_format:
-                mock_repo = _make_repo_mock()
-                mock_repo.view.return_value = simple_conversation
-                mock_get_repo.return_value = mock_repo
-                mock_format.return_value = "# Content"
-
-                server = _build_server()
-                _invoke_surface(server._tool_manager._tools["export_conversation"].fn,
-                    id="test:conv-123", format="invalid_format"
-                )
-
-                # Verify markdown was used as fallback
-                call_args = mock_format.call_args
-                assert call_args[0][1] == "markdown"
-
-
-# =============================================================================
-# New Resource Tests
-# =============================================================================
-
-
-class TestNewResources:
-    """Tests for new resources."""
-
-    def test_tags_resource(self):
-        """polylogue://tags resource returns tag counts."""
-        from polylogue.mcp.server import _build_server
-
-        with patch("polylogue.mcp.server._get_repo") as mock_get_repo:
-            mock_repo = _make_repo_mock()
-            mock_repo.list_tags.return_value = {"feature": 10, "bug": 5}
-            mock_get_repo.return_value = mock_repo
-
-            server = _build_server()
-            result = _invoke_surface(server._resource_manager._resources["polylogue://tags"].fn, )
-
-            parsed = json.loads(result)
-            assert parsed["feature"] == 10
-            assert parsed["bug"] == 5
-
-    def test_health_resource(self):
-        """polylogue://health resource returns health status."""
-        from polylogue.mcp.server import _build_server
-
-        mock_check = MagicMock()
-        mock_check.name = "database"
-        mock_check.status.value = "ok"
-
-        mock_report = MagicMock()
-        mock_report.checks = [mock_check]
-        mock_report.summary = "All systems operational"
-
-        with patch("polylogue.mcp.server._get_config") as mock_get_config:
-            with patch("polylogue.health.get_health") as mock_get_health:
-                mock_get_config.return_value = MagicMock()
-                mock_get_health.return_value = mock_report
-
-                server = _build_server()
-                result = _invoke_surface(server._resource_manager._resources["polylogue://health"].fn, )
-
-                parsed = json.loads(result)
-                assert "checks" in parsed
-                assert parsed["summary"] == "All systems operational"
-
-
-# =============================================================================
-# New Prompt Tests
-# =============================================================================
-
-
-class TestNewPrompts:
-    """Tests for new prompts."""
-
-    def test_compare_conversations_prompt(self, simple_conversation):
-        """compare_conversations prompt returns comparison text."""
-        from polylogue.mcp.server import _build_server
-
-        with patch("polylogue.mcp.server._get_repo") as mock_get_repo:
-            mock_repo = _make_repo_mock()
-            mock_repo.view.side_effect = [simple_conversation, simple_conversation]
-            mock_get_repo.return_value = mock_repo
-
-            server = _build_server()
-            result = _invoke_surface(server._prompt_manager._prompts["compare_conversations"].fn,
-                id1="test:conv-1", id2="test:conv-2"
-            )
-
-            assert isinstance(result, str)
-            assert "Compare" in result
-            assert "Conversation 1" in result
-            assert "Conversation 2" in result
-
-    @pytest.mark.asyncio
-    async def test_extract_patterns_prompt(self, simple_conversation):
-        """extract_patterns prompt returns pattern analysis text."""
-        from polylogue.mcp.server import _build_server
-
-        with patch("polylogue.mcp.server._get_repo") as mock_get_repo:
-            with patch("polylogue.lib.filters.ConversationFilter") as MockFilter:
-                mock_repo = _make_repo_mock()
-                mock_get_repo.return_value = mock_repo
-                MockFilter.return_value = make_mock_filter(results=[simple_conversation])
-
-                server = _build_server()
-                result = await _invoke_surface_async(server._prompt_manager._prompts["extract_patterns"].fn, )
-
-                assert isinstance(result, str)
-                assert "patterns" in result.lower()
-
-
-# =============================================================================
-# Error Handling Tests
-# =============================================================================
-
-
-class TestSafeCall:
-    """Tests for _safe_call error wrapper."""
-
-    def test_safe_call_returns_result(self):
-        """_safe_call returns result when fn succeeds."""
-        from polylogue.mcp.server import _safe_call
-
-        def success_fn():
-            return "ok"
-
-        result = _safe_call("test_fn", success_fn)
-        assert result == "ok"
-
-    def test_safe_call_catches_exception(self):
-        """_safe_call catches exceptions and returns error JSON."""
-        from polylogue.mcp.server import _safe_call
-
-        def failing_fn():
-            raise ValueError("Something went wrong")
-
-        result = _safe_call("test_fn", failing_fn)
-        parsed = json.loads(result)
-        assert "error" in parsed
-        assert "Something went wrong" in parsed["error"]
-        assert parsed["tool"] == "test_fn"
