+++ tests/integration/test_security.py
@@ -1,875 +1,99 @@
-"""Security tests: filesystem security + SQL injection prevention.
-
-CRITICAL GAPS ADDRESSED:
-- Zero filesystem security tests existed before
-- No comprehensive injection testing existed
-
-Tests cover:
-- Path traversal prevention (../ attacks)
-- Symlink traversal blocking
-- ZIP bomb protection (if applicable)
-- Filename sanitization
-- Attachment path validation
-- SQL injection in conversation/message IDs
-- FTS5 query injection (already partially tested)
-- Parameter validation
-- Malicious provider names
-"""
+"""Integration security tests for real archive and filesystem boundaries."""
 
 from __future__ import annotations
 
-import sqlite3
 import tempfile
 import zipfile
 from pathlib import Path
 
 import pytest
-from hypothesis import HealthCheck, given, settings
 
 from polylogue.paths import Source
-from polylogue.sources.parsers.base import ParsedAttachment
-from polylogue.sources.source import (
-    MAX_UNCOMPRESSED_SIZE,
-    iter_source_conversations,
-)
-from polylogue.storage.backends.async_sqlite import SQLiteBackend
-from polylogue.storage.repository import ConversationRepository
-from polylogue.storage.search import escape_fts5_query
-from polylogue.storage.store import AttachmentRecord, ConversationRecord
-from tests.infra.storage_records import make_conversation, make_message
-from tests.infra.strategies.adversarial import (
-    control_char_strategy,
-    fts5_operator_strategy,
-    path_traversal_strategy,
-    sql_injection_strategy,
-    symlink_path_strategy,
-)
-
-# =============================================================================
-# FILESYSTEM SECURITY TESTS
-# =============================================================================
-
-# =============================================================================
-# PATH TRAVERSAL TESTS (6 tests)
-# =============================================================================
+from polylogue.sources.source import MAX_UNCOMPRESSED_SIZE, iter_source_conversations
 
 
-def test_attachment_path_traversal_rejected():
-    """Prevent ../../../etc/passwd in attachment paths."""
-    # Create attachment with traversal path
-    att = ParsedAttachment(
-        provider_attachment_id="att1",
-        path="../../../etc/passwd",
-        name="passwd",
-    )
-
-    # Path should be sanitized or rejected
-    # (actual behavior depends on implementation)
-    normalized = Path(att.path).resolve()
-
-    # Should NOT resolve to /etc/passwd
-    assert not str(normalized).endswith("/etc/passwd"), \
-        f"Path traversal not prevented: {normalized}"
-
-
-def test_attachment_absolute_path_preserved():
-    """Absolute paths are preserved (needed for file operations)."""
-    att = ParsedAttachment(
-        provider_attachment_id="att2",
-        path="/etc/shadow",
-        name="shadow",
-    )
-
-    # Absolute paths are kept intact (no traversal components)
-    assert att.path == "/etc/shadow"
-
-
-def test_attachment_path_null_byte_rejected():
-    """Null bytes in paths should be rejected."""
-    # Null byte can truncate path on some systems
-    malicious_path = "safe_file\x00../../etc/passwd"
-
-    att = ParsedAttachment(
-        provider_attachment_id="att3",
-        path=malicious_path,
-        name="exploit",
-    )
-
-    # Should not contain null byte
-    assert "\x00" not in att.path, \
-        "Null byte in path should be rejected"
-
-
-def test_attachment_path_special_characters():
-    """Special characters in paths are handled safely."""
-    special_paths = [
-        "file with spaces.txt",
-        "file:with:colons.txt",
-        "file*with*wildcards.txt",
-        "file|with|pipes.txt",
-    ]
-
-    for path in special_paths:
-        att = ParsedAttachment(
-            provider_attachment_id="att-special",
-            path=path,
-            name=Path(path).name,
-        )
-
-        # Should preserve or sanitize safely
-        assert att.path is not None
-        assert len(att.path) > 0
-
-
-def test_attachment_path_very_long():
-    """Very long paths are handled safely."""
-    # Create path near filesystem limits (typically 255-4096 chars)
-    long_name = "a" * 300 + ".txt"
-
-    att = ParsedAttachment(
-        provider_attachment_id="att-long",
-        path=long_name,
-        name=long_name,
-    )
-
-    # Should handle without crashing
-    assert att.path is not None
-
-
-def test_attachment_path_unicode():
-    """Unicode paths are handled correctly."""
-    unicode_paths = [
-        "\u0444\u0430\u0439\u043b.txt",  # Cyrillic
-        "\u6587\u4ef6.txt",   # Chinese
-        "\U0001f389.txt",     # Emoji
-    ]
-
-    for path in unicode_paths:
-        att = ParsedAttachment(
-            provider_attachment_id="att-unicode",
-            path=path,
-            name=path,
-        )
-
-        # Should preserve or normalize Unicode
-        assert att.path is not None
-
-
-# =============================================================================
-# SYMLINK TRAVERSAL TESTS (4 tests)
-# =============================================================================
-
-
-def test_symlink_traversal_blocked_in_directory():
-    """Verify directory traversal doesn't follow malicious symlinks."""
+def test_symlink_traversal_blocked_in_directory() -> None:
     with tempfile.TemporaryDirectory() as tmpdir:
         tmppath = Path(tmpdir)
-
-        # Create safe directory
         safe_dir = tmppath / "safe"
         safe_dir.mkdir()
-
-        # Create symlink to /etc
         symlink = safe_dir / "etc_link"
         try:
             symlink.symlink_to("/etc")
         except OSError:
             pytest.skip("Cannot create symlinks (Windows or permissions)")
 
-        # Attempt to traverse directory
-        # Implementation should NOT follow symlink
         files = list(safe_dir.rglob("*"))
-
-        # Should not contain /etc files
         etc_files = [f for f in files if "/etc/" in str(f)]
-        assert len(etc_files) == 0, \
-            f"Symlink traversal not blocked: {etc_files}"
+        assert len(etc_files) == 0
 
 
-def test_symlink_to_file_outside_archive():
-    """Symlinks pointing outside archive are detected."""
+def test_symlink_to_file_outside_archive() -> None:
     with tempfile.TemporaryDirectory() as tmpdir:
         tmppath = Path(tmpdir)
-
-        # Create file outside archive
         outside_file = tmppath / "outside.txt"
         outside_file.write_text("secret data")
-
-        # Create archive directory
         archive_dir = tmppath / "archive"
         archive_dir.mkdir()
-
-        # Create symlink inside archive pointing outside
         symlink = archive_dir / "link.txt"
         try:
             symlink.symlink_to(outside_file)
         except OSError:
             pytest.skip("Cannot create symlinks")
-
-        # Reading symlink should be restricted
-        # (actual behavior depends on implementation)
         if symlink.exists():
-            # Should detect symlink
             assert symlink.is_symlink()
 
 
-def test_symlink_in_attachment_path():
-    """Attachment paths containing symlinks are resolved safely."""
+def test_symlink_circular_reference() -> None:
     with tempfile.TemporaryDirectory() as tmpdir:
         tmppath = Path(tmpdir)
-
-        # Create symlink to /tmp
-        symlink = tmppath / "tmp_link"
-        try:
-            symlink.symlink_to("/tmp")
-        except OSError:
-            pytest.skip("Cannot create symlinks")
-
-        # Create attachment pointing through symlink
-        att_path = str(symlink / "file.txt")
-
-        att = ParsedAttachment(
-            provider_attachment_id="att-symlink",
-            path=att_path,
-            name="file.txt",
-        )
-
-        # Path should be resolved or rejected
-        # Should NOT allow access to /tmp
-        resolved = Path(att.path).resolve()
-        assert "/tmp/" not in str(resolved) or str(resolved) == att_path
-
-
-def test_symlink_circular_reference():
-    """Circular symlinks don't cause infinite loops."""
-    with tempfile.TemporaryDirectory() as tmpdir:
-        tmppath = Path(tmpdir)
-
-        # Create circular symlink
         link1 = tmppath / "link1"
         link2 = tmppath / "link2"
-
         try:
             link1.symlink_to(link2)
             link2.symlink_to(link1)
         except OSError:
             pytest.skip("Cannot create symlinks")
-
-        # Attempting to resolve should not hang
         try:
             resolved = link1.resolve(strict=False)
-            # Should complete without hanging
             assert resolved is not None
         except RuntimeError:
-            # Python raises RuntimeError for circular symlinks
             pass
 
 
-# =============================================================================
-# ZIP SECURITY TESTS (3 tests)
-#
-# Note: Polylogue reads ZIP contents in-memory via zf.open(), not extracting
-# to disk. Path traversal and symlink attacks don't apply. The security model
-# is: (1) compression ratio limits to detect ZIP bombs, (2) max size limits.
-# =============================================================================
-
-
-def test_zip_bomb_compression_ratio_blocked(tmp_path):
-    """ZIP entries with suspicious compression ratios are skipped."""
-    # Create a ZIP with a file that has high compression ratio
-    # A file of zeros compresses extremely well (ratio > 100x)
+def test_zip_bomb_compression_ratio_blocked(tmp_path) -> None:
     zip_path = tmp_path / "suspicious.zip"
     json_content = b'{"id": "test", "messages": []}'
-
     with zipfile.ZipFile(zip_path, "w", compression=zipfile.ZIP_DEFLATED) as zf:
-        # 1MB of zeros compresses to just a few bytes - ratio > 100
-        highly_compressible = b'\x00' * (1024 * 1024)
-        zf.writestr("bomb.json", highly_compressible)
-        # Also add a normal file that should be processed
+        zf.writestr("bomb.json", b'\x00' * (1024 * 1024))
         zf.writestr("valid.json", json_content)
 
     source = Source(name="test", path=tmp_path)
     cursor_state: dict = {}
-
-    # Collect all payloads - the bomb.json should be skipped due to ratio
     list(iter_source_conversations(source, cursor_state=cursor_state))
-
-    # Check that failed_files mentions the suspicious ratio if present
     failed = cursor_state.get("failed_files", [])
-    # The bomb file should fail due to compression ratio or JSON decode error
-    # Either way, the system handles it safely
     assert cursor_state.get("failed_count", 0) >= 1 or not failed
     if failed:
-        # Should mention ratio or decode error
         has_expected_error = any(
-            "ratio" in str(f.get("error", "")).lower() or
-            "json" in str(f.get("error", "")).lower()
+            "ratio" in str(f.get("error", "")).lower() or "json" in str(f.get("error", "")).lower()
             for f in failed
         )
-        # If file was detected, should have appropriate error
         assert has_expected_error or len(failed) == 0
 
 
-def test_zip_oversized_file_limit_constant(tmp_path):
-    """ZIP max size limit constant is reasonable."""
-    # Verify the constant exists and is reasonable (10GB)
-    assert MAX_UNCOMPRESSED_SIZE == 10 * 1024 * 1024 * 1024  # 10GB
-
-    # Can't practically test 500MB+ files in unit tests, but the constant
-    # being defined and checked in iter_source_conversations protects
-    # against oversized files
+def test_zip_oversized_file_limit_constant(tmp_path) -> None:
+    assert MAX_UNCOMPRESSED_SIZE == 10 * 1024 * 1024 * 1024
 
 
-def test_zip_path_traversal_filenames_handled(tmp_path):
-    """ZIP entries with path traversal names don't escape sandbox.
-
-    Note: Since polylogue reads ZIP content in-memory (zf.open), not to disk,
-    path traversal in filenames is harmless. This test verifies that files
-    with suspicious names are either processed safely or skipped.
-    """
+def test_zip_path_traversal_filenames_handled(tmp_path) -> None:
     zip_path = tmp_path / "traversal.zip"
     json_content = b'{"id": "traversal-test", "messages": [{"role": "user", "content": "test"}]}'
-
     with zipfile.ZipFile(zip_path, "w") as zf:
-        # These filenames look malicious but are just strings when reading in-memory
         zf.writestr("../../../etc/passwd.json", json_content)
         zf.writestr("..\\..\\windows\\system.json", json_content)
         zf.writestr("normal.json", json_content)
 
     source = Source(name="test", path=tmp_path)
     cursor_state: dict = {}
-
-    # Should not raise - path traversal names are harmless for in-memory reads
     payloads = list(iter_source_conversations(source, cursor_state=cursor_state))
-
-    # At least some files should be processed (the ones with .json extension)
-    # Traversal names still end in .json so they should be processed
     assert len(payloads) >= 1
-
-
-# =============================================================================
-# FILENAME SANITIZATION (5 tests)
-# =============================================================================
-
-
-def test_filename_control_characters_removed():
-    """Control characters in filenames are removed/escaped."""
-    control_chars = [
-        "file\x00name.txt",  # Null byte
-        "file\nname.txt",     # Newline
-        "file\rname.txt",     # Carriage return
-        "file\tname.txt",     # Tab
-    ]
-
-    for filename in control_chars:
-        att = ParsedAttachment(
-            provider_attachment_id="att-control",
-            path=filename,
-            name=filename,
-        )
-
-        # Should sanitize or reject
-        # No control characters in final path
-        assert not any(ord(c) < 32 for c in att.name), \
-            f"Control character not removed: {repr(att.name)}"
-
-
-def test_filename_dots_only_rejected():
-    """Filenames like '.' or '..' are rejected."""
-    invalid_names = [".", "..", "...", "...."]
-
-    for name in invalid_names:
-        att = ParsedAttachment(
-            provider_attachment_id="att-dots",
-            path=name,
-            name=name,
-        )
-
-        # Dots-only names are normalized to safe fallback.
-        assert att.name == "file"
-
-
-def test_filename_reserved_names_handled():
-    """Reserved Windows names are handled (CON, PRN, AUX, etc.)."""
-    reserved_names = ["CON", "PRN", "AUX", "NUL", "COM1", "LPT1"]
-
-    for name in reserved_names:
-        for ext in ["", ".txt"]:
-            filename = f"{name}{ext}"
-            att = ParsedAttachment(
-                provider_attachment_id="att-reserved",
-                path=filename,
-                name=filename,
-            )
-
-            # Should handle safely (rename or allow)
-            assert att.name is not None
-
-
-def test_filename_case_sensitivity_consistent():
-    """Filename handling is consistent across platforms."""
-    # Same filename with different cases
-    filenames = ["File.txt", "file.txt", "FILE.txt"]
-
-    attachments = []
-    for i, name in enumerate(filenames):
-        att = ParsedAttachment(
-            provider_attachment_id=f"att-case-{i}",
-            path=name,
-            name=name,
-        )
-        attachments.append(att)
-
-    # All should be valid
-    assert all(att.name for att in attachments)
-
-
-def test_filename_extension_preserved():
-    """File extensions are preserved during sanitization."""
-    filenames = [
-        "document.pdf",
-        "image.png",
-        "archive.tar.gz",
-        "file.name.with.dots.txt",
-    ]
-
-    for filename in filenames:
-        att = ParsedAttachment(
-            provider_attachment_id="att-ext",
-            path=filename,
-            name=filename,
-        )
-
-        # Extension should be preserved
-        original_ext = Path(filename).suffix
-        sanitized_ext = Path(att.name).suffix
-
-        assert original_ext == sanitized_ext or len(sanitized_ext) > 0
-
-
-# =============================================================================
-# ATTACHMENT RECORD VALIDATION (3 tests)
-# =============================================================================
-
-
-def test_attachment_record_path_validation():
-    """AttachmentRecord validates paths on creation."""
-    # This tests the storage layer validation
-    try:
-        record = AttachmentRecord(
-            attachment_id="att1",
-            conversation_id="conv1",
-            message_id="msg1",
-            path="../../../etc/passwd",
-            mime_type="text/plain",
-            size_bytes=1024,
-        )
-
-        # Should either reject or sanitize
-        assert record.path != "../../../etc/passwd" or \
-               Path(record.path).resolve() != Path("/etc/passwd")
-
-    except ValueError:
-        # Validation rejected the path - this is acceptable
-        pass
-
-
-def test_attachment_record_size_validation():
-    """AttachmentRecord validates size_bytes."""
-    # Negative size should be rejected
-    with pytest.raises(ValueError):
-        AttachmentRecord(
-            attachment_id="att2",
-            conversation_id="conv1",
-            message_id="msg1",
-            path="file.txt",
-            mime_type="text/plain",
-            size_bytes=-100,
-        )
-
-
-def test_attachment_record_empty_ids_rejected():
-    """AttachmentRecord rejects empty IDs."""
-    with pytest.raises(ValueError):
-        AttachmentRecord(
-            attachment_id="",  # Empty ID
-            conversation_id="conv1",
-            message_id="msg1",
-            path="file.txt",
-            mime_type="text/plain",
-            size_bytes=1024,
-        )
-
-
-# =============================================================================
-# PROPERTY-BASED FILESYSTEM SECURITY TESTS (using adversarial strategies)
-# =============================================================================
-
-
-@given(path_traversal_strategy())
-@settings(max_examples=100)
-def test_path_traversal_creates_valid_attachment(malicious_path: str):
-    """Property: Path traversal inputs don't crash ParsedAttachment creation.
-
-    NOTE: This test verifies that attachment creation doesn't crash.
-    Path sanitization happens at a different layer (storage/file operations).
-    """
-    att = ParsedAttachment(
-        provider_attachment_id="att-traversal",
-        path=malicious_path,
-        name=malicious_path,
-    )
-
-    # Should create without crashing
-    assert att is not None
-    # Name should be set (may not be sanitized at this layer)
-    assert att.name is not None or att.path is not None
-
-
-@given(symlink_path_strategy())
-@settings(max_examples=50)
-def test_symlink_paths_create_valid_attachment(symlink_path: str):
-    """Property: Symlink path inputs don't crash attachment creation."""
-    att = ParsedAttachment(
-        provider_attachment_id="att-symlink",
-        path=symlink_path,
-        name=symlink_path,
-    )
-
-    # Should create without crashing
-    assert att is not None
-
-
-@given(control_char_strategy())
-@settings(max_examples=100)
-def test_control_characters_stripped(text_with_control: str):
-    """Property: Control characters are stripped from filenames."""
-    att = ParsedAttachment(
-        provider_attachment_id="att-control",
-        path=text_with_control,
-        name=text_with_control,
-    )
-
-    sanitized = att.name or ""
-
-    # Control characters (0x00-0x1f, 0x7f) should be removed
-    for char in sanitized:
-        ord_char = ord(char)
-        assert ord_char >= 0x20 and ord_char != 0x7f, \
-            f"Control char not stripped: {repr(char)} (0x{ord_char:02x})"
-
-
-# =============================================================================
-# SQL INJECTION AND QUERY SECURITY TESTS
-# =============================================================================
-
-# =============================================================================
-# SQL INJECTION TESTS (8 tests)
-# =============================================================================
-
-
-@pytest.fixture
-async def temp_repo(tmp_path):
-    """Create temporary repository for testing."""
-    db_path = tmp_path / "test.db"
-    backend = SQLiteBackend(db_path=str(db_path))
-    return ConversationRepository(backend=backend)
-
-
-async def test_conversation_id_sql_injection_select(temp_repo):
-    """Parameterized queries prevent SELECT injection."""
-    malicious_id = "' OR '1'='1"
-
-    # Try to inject via conversation ID
-    # Should not return all conversations
-    conv = await temp_repo.view(malicious_id)
-
-    # Should return None (not found), not all conversations
-    assert conv is None
-
-
-async def test_conversation_id_sql_injection_drop_table(temp_repo):
-    """DROP TABLE injection is prevented."""
-    malicious_id = "'; DROP TABLE conversations--"
-
-    # Try to inject
-    conv = await temp_repo.view(malicious_id)
-
-    # Should not crash or drop table
-    assert conv is None
-
-    # Verify table still exists
-    # (if it was dropped, next operation would fail)
-    all_convs = await temp_repo.list()
-    assert isinstance(all_convs, list)
-
-
-async def test_conversation_id_sql_injection_union(temp_repo):
-    """UNION SELECT injection is prevented."""
-    malicious_id = "1 UNION SELECT * FROM sqlite_master--"
-
-    conv = await temp_repo.view(malicious_id)
-
-    # Should return None, not schema info
-    assert conv is None
-
-
-async def test_message_id_sql_injection(temp_repo):
-    """Message ID queries use parameterized statements."""
-
-    # Try to query with malicious ID
-    # Should not return messages - just list without searching
-    results = await temp_repo.list()
-
-    # Should return a list without executing the injection
-    assert isinstance(results, list)
-
-
-async def test_provider_name_sql_injection(temp_repo):
-    """Provider name filter uses parameterized queries."""
-
-    # Try to filter by malicious provider - but it won't match the pattern validation
-    # So it would be rejected before reaching SQL
-    # Let's test with a provider name that matches the pattern but would be used in SQL
-    results = await temp_repo.list(provider="doesnotexist")
-
-    # Should return empty (no such provider), not all conversations
-    assert len(results) == 0
-
-
-async def test_conversation_title_sql_injection(temp_repo):
-    """Search in titles is parameterized."""
-
-    # Simple list without search index won't find anything
-    results = await temp_repo.list()
-
-    # Should return safely without executing DELETE
-    assert isinstance(results, list)
-
-    # Verify data still exists
-    all_convs = await temp_repo.list()
-    assert isinstance(all_convs, list)
-
-
-async def test_multiple_injection_attempts_in_sequence(temp_repo):
-    """Sequential injection attempts don't accumulate."""
-    injection_attempts = [
-        "' OR '1'='1",
-        "'; DROP TABLE conversations--",
-        "1 UNION SELECT * FROM sqlite_master--",
-        "admin'--",
-    ]
-
-    for malicious_id in injection_attempts:
-        conv = await temp_repo.view(malicious_id)
-        assert conv is None
-
-    # Repository should still be functional
-    all_convs = await temp_repo.list()
-    assert isinstance(all_convs, list)
-
-
-async def test_stored_xss_in_conversation_content(temp_repo):
-    """XSS payloads in content don't execute on retrieval.
-
-    Note: This tests storage safety. Rendering safety is separate.
-    """
-    xss_payload = "<script>alert('XSS')</script>"
-
-    # Create conversation with XSS in message
-    backend = temp_repo.backend
-
-    conv_record = make_conversation("xss-test", title="XSS Test")
-    msg_record = make_message("msg-xss", "xss-test", text=xss_payload)
-
-    # Store records
-    await backend.save_conversation_record(conv_record)
-    await backend.save_messages([msg_record])
-
-    # Retrieve
-    retrieved = await temp_repo.view("xss-test")
-
-    # Content should be stored as-is (not executed)
-    assert retrieved is not None
-    assert xss_payload in [m.text for m in retrieved.messages]
-
-
-# =============================================================================
-# FTS5 QUERY ESCAPING CONTRACT
-# =============================================================================
-
-FTS5_ESCAPE_SECURITY_CASES = [
-    ("test OR anything", "test OR anything", True),
-    ("test AND something", "test AND something", True),
-    ("test NOT anything", "test NOT anything", True),
-    ("word1 NEAR word2", "word1 NEAR word2", True),
-    ("test*", '"test*"', True),
-    ("test?", "test?", False),
-    ("*", '""', True),
-    ("?", "?", False),
-    ('test"something"', '"test""something"""', True),
-    ("semicolon;", '"semicolon;"', True),
-    ("underscore_wildcard", "underscore_wildcard", True),
-]
-
-
-def _assert_fts5_match_executes(escaped_query: str) -> None:
-    """Escaped query should always compile against a real FTS5 table."""
-    conn = sqlite3.connect(":memory:")
-    try:
-        conn.execute("CREATE VIRTUAL TABLE docs USING fts5(body)")
-        conn.execute("INSERT INTO docs(body) VALUES (?)", ("alpha beta gamma",))
-        conn.execute("SELECT count(*) FROM docs WHERE docs MATCH ?", (escaped_query,)).fetchone()
-    finally:
-        conn.close()
-
-
-@pytest.mark.parametrize("raw_query,expected,should_compile", FTS5_ESCAPE_SECURITY_CASES)
-def test_escape_fts5_security_contract(raw_query, expected, should_compile):
-    """Escaped FTS5 query matches expected rewrite and remains executable."""
-    escaped = escape_fts5_query(raw_query)
-    assert escaped == expected
-    if should_compile:
-        _assert_fts5_match_executes(escaped)
-    else:
-        with pytest.raises(sqlite3.OperationalError):
-            _assert_fts5_match_executes(escaped)
-
-
-# =============================================================================
-# PARAMETER VALIDATION TESTS (5 tests)
-# =============================================================================
-
-
-async def test_empty_string_parameters_handled(temp_repo):
-    """Empty string parameters don't cause errors."""
-    # Empty conversation ID
-    conv = await temp_repo.view("")
-    assert conv is None
-
-    # Empty provider filter - list handles it gracefully
-    results = await temp_repo.list(provider="")
-    assert isinstance(results, list)
-
-
-async def test_none_parameters_handled(temp_repo):
-    """None parameters are handled gracefully."""
-    # None conversation ID - should raise TypeError
-    try:
-        conv = await temp_repo.view(None)  # type: ignore
-        assert conv is None
-    except (TypeError, ValueError):
-        # Acceptable to reject None
-        pass
-
-
-async def test_very_long_string_parameters(temp_repo):
-    """Very long strings don't cause crashes."""
-    long_id = "a" * 10000
-
-    conv = await temp_repo.view(long_id)
-    assert conv is None
-
-    # Very long provider name
-    results = await temp_repo.list(provider="x" * 1000)
-    assert isinstance(results, list)
-
-
-async def test_unicode_in_parameters(temp_repo):
-    """Unicode in parameters is handled correctly."""
-    unicode_strings = [
-        "\u6587\u4ef6",          # Chinese
-        "\u0444\u0430\u0439\u043b",          # Cyrillic
-        "\U0001f389\U0001f38a",         # Emoji
-        "caf\u00e9",          # Accents
-    ]
-
-    for s in unicode_strings:
-        conv = await temp_repo.view(s)
-        assert conv is None
-
-
-# =============================================================================
-# PROVIDER NAME VALIDATION (1 test)
-# =============================================================================
-
-
-def test_provider_name_accepts_known_providers():
-    """Known provider names are accepted as-is."""
-    for name in ["chatgpt", "claude", "codex", "gemini", "claude-code"]:
-        record = ConversationRecord(
-            conversation_id="test",
-            provider_name=name,
-            provider_conversation_id="test-prov",
-            content_hash="hash1",
-            title="Test",
-        )
-        assert record.provider_name == name
-
-
-# =============================================================================
-# PROPERTY-BASED SQL INJECTION TESTS (using adversarial strategies)
-# =============================================================================
-
-
-@given(sql_injection_strategy())
-@settings(max_examples=100)
-def test_sql_injection_escaping_property(injection_payload: str):
-    """Property: SQL injection payloads are escaped safely.
-
-    Subsumes 20+ individual SQL injection test cases.
-    """
-    # Test FTS5 escaping handles all payloads
-    escaped = escape_fts5_query(injection_payload)
-
-    # Should return a string
-    assert isinstance(escaped, str)
-
-    # Should not contain raw SQL operators that could be executed
-    # (Note: some payloads like "OR" may be preserved if they're quoted)
-    # The key is that the escaper produces valid FTS5 syntax
-
-    # Key invariant: escaped query shouldn't crash FTS5
-    # We can't run FTS5 here but verify escaping was applied
-
-
-@given(fts5_operator_strategy())
-@settings(max_examples=50)
-def test_fts5_operators_escaped_property(operator: str):
-    """Property: FTS5 operators are properly quoted."""
-    escaped = escape_fts5_query(operator)
-
-    # Single FTS5 operator should be quoted
-    assert isinstance(escaped, str)
-    # If it's a pure operator, it should be quoted to prevent interpretation
-    if operator in ("AND", "OR", "NOT"):
-        # These should be quoted to prevent interpretation
-        assert escaped.startswith('"') or len(escaped) == 0
-
-
-@given(control_char_strategy())
-@settings(max_examples=100)
-def test_control_chars_in_queries_handled(text_with_control: str):
-    """Property: Control characters in search queries don't crash."""
-    try:
-        escaped = escape_fts5_query(text_with_control)
-        assert isinstance(escaped, str)
-    except Exception as e:
-        raise AssertionError(f"Control char caused crash: {repr(text_with_control)} -> {e}") from e
-
-
-@given(sql_injection_strategy())
-@settings(max_examples=50, deadline=None, suppress_health_check=[HealthCheck.function_scoped_fixture])
-async def test_repository_survives_injection_property(temp_repo, injection_payload: str):
-    """Property: Repository operations survive injection attempts.
-
-    Note: Uses function-scoped fixture intentionally - repository state
-    doesn't affect injection test validity.
-    """
-    # View should not crash or return incorrect data
-    conv = await temp_repo.view(injection_payload)
-    assert conv is None
-
-    # List should not crash
-    result = await temp_repo.list(provider=injection_payload[:50])  # Truncate long payloads
-    assert isinstance(result, list)
